chore: [app-router-migration 8.6] reorganize future pages file structure (#12988)
* make no-meeting-found page use ssr

* remove-route-groups

* add LayoutHOC

* fix

* fix

* ensure proper types for withLayout function

---------
