+++ apps/web/app/_types.ts
@@ -1,3 +1,14 @@
 export type Params = {
   [param: string]: string | string[] | undefined;
 };
+
+export type SearchParams = {
+  [param: string]: string | string[] | undefined;
+};
+
+export type PageProps = {
+  params: Params;
+  searchParams: SearchParams;
+};
+
+export type LayoutProps = { params: Params; children: React.ReactElement };
