--- apps/web/app/future/(shared-page-wrapper)/(admin-layout)/layout.tsx
@@ -1,20 +1,0 @@
-import { headers } from "next/headers";
-import { type ReactElement } from "react";
-
-import PageWrapper from "@components/PageWrapperAppDir";
-import { getLayout } from "@components/auth/layouts/AdminLayoutAppDir";
-
-type WrapperWithLayoutProps = {
-  children: ReactElement;
-};
-
-export default async function WrapperWithLayout({ children }: WrapperWithLayoutProps) {
-  const h = headers();
-  const nonce = h.get("x-nonce") ?? undefined;
-
-  return (
-    <PageWrapper getLayout={getLayout} requiresLicense={false} nonce={nonce} themeBasis={null}>
-      {children}
-    </PageWrapper>
-  );
-}
