--- apps/web/app/future/(shared-page-wrapper)/(no-layout)/video/no-meeting-found/page.tsx
@@ -1,10 +1,0 @@
-import Page from "@pages/video/no-meeting-found";
-import { _generateMetadata } from "app/_utils";
-
-export const generateMetadata = async () =>
-  await _generateMetadata(
-    () => "",
-    () => ""
-  );
-
-export default Page;
