+++ apps/web/app/future/apps/[slug]/page.tsx
