+++ apps/web/app/future/apps/[slug]/setup/page.tsx
