+++ apps/web/app/future/settings/admin/apps/[category]/page.tsx
