+++ apps/web/app/future/settings/admin/oAuth/oAuthView/layout.tsx
@@ -1,0 +1,3 @@
+import { WithLayout } from "app/layoutHOC";
+
+export default WithLayout({ getLayout: null })<"L">;
